cell_cols {cellranger}	R Documentation

Specify cell limits only for columns

Description

How does this differ from cell_limits? Two ways. First, the input can have length greater than 2, i.e. the columns can be specified as 1:n. If the length is greater than 2, both the min and max are taken with NA.rm = TRUE. Note it is not possible to request non-contiguous columns, i.e. columns 1, 2, and 5. In this case, the requested columns will run from the minimum of 1 to the maximum of 5. Second, the input can be given in the letter-based format spreadsheets use to label columns.

Usage

cell_cols(x)

Arguments

x
vector of column limits; if character, converted to numeric; if length greater than two, min and max will be taken with NA.rm = TRUE

Value

a cell_limits object

Examples

cell_cols(c(NA, 3))
cell_cols(c(7, NA))
cell_cols(4:16)
cell_cols(c(3, NA, 10))

cell_cols("C:G")
cell_cols(c("B", NA))
cell_cols(LETTERS)

[Package cellranger version 1.1.0 Index]